Added submodule, changed colorscheme, rc
Added emmet submodule for html coding.
Added matchit.vim to runtimte (for better matching using %)
Changed colorscheme to gruvbox. Lovin' it.
Added highlighting when navigating search results.
Changed nerdtreetabs options

Added options for taglist.
Need to figure out a way to add it to the repo. It doesn't have a
git repo to add as a submodule.